SPARC64: implement custom PCI bus scan for sabre
The sabre on-board devices are located at devfn (1,1) which means that in a
standard PCI bus scan the bus at devfn (1, 0) is enumerated first. This
doesn't play well with OpenBIOS which assumes that the ebus (with legacy
ioports) is mapped starting from io_base == 0x0.

Perform a custom scan to ensure that OpenBIOS doesn't crash when additional
devices are added to the bus at devfn (1, 0).

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>