pci: rename SUN4U_INTERRUPT to SUN4U_PCIAINTERRUPT
The current interrupt mapping is only correct for PCI bus A (QEMU
currently alters its PCI mappings to handle that PCI devices are
plugged into the PCI root bus and not behind a Simba bridge).

Add a comment to point out that the existing mapping is deprecated
and should be removed once the QEMU PCI topology has been updated.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>