.github/workflows: add release.yml for generating an OpenBIOS release
This is a GitHub push action that builds OpenBIOS for all of the currently
supported architectures (amd64, sparc32, sparc64, ppc and x86) and a new
GitHub release consisting of an output zip file containing debug and release
binaries, along with the source in .tar.gz and .zip formats.

A release build is triggered by pushing a tag beginning with "v" indicating a
version number in contrast to pushing a branch.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>